Add more builder composition options - remove singleton runner
