+++ SharperUniverse.sln
@@ -9,6 +9,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ExampleECSUsageConsole", "E
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SharperUniverse.Tests", "SharperUniverse.Tests\SharperUniverse.Tests.csproj", "{58928961-65C1-4790-82A3-1B5BAE326378}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SharperUniverse.Example.GuessingGame", "SharperUniverse.Example.GuessingGame\SharperUniverse.Example.GuessingGame.csproj", "{DE9D4D1D-40CC-4467-A8AE-9630BE0444EE}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -27,6 +29,10 @@ Global
 		{58928961-65C1-4790-82A3-1B5BAE326378}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{58928961-65C1-4790-82A3-1B5BAE326378}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{58928961-65C1-4790-82A3-1B5BAE326378}.Release|Any CPU.Build.0 = Release|Any CPU
+		{DE9D4D1D-40CC-4467-A8AE-9630BE0444EE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{DE9D4D1D-40CC-4467-A8AE-9630BE0444EE}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{DE9D4D1D-40CC-4467-A8AE-9630BE0444EE}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{DE9D4D1D-40CC-4467-A8AE-9630BE0444EE}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
